YT-18571: Extract and refactor execution stacks commit_hash:856b6638d9642f4711f5058d77bd58dd78accb8a
