+++ yt/yt/core/concurrency/fiber.h
@@ -127,7 +127,7 @@ public:
     static void ReadFibers(TFunctionView<void(TFiberList&)> callback);
 
 private:
-    const std::shared_ptr<TExecutionStack> Stack_;
+    const std::shared_ptr<NThreading::TExecutionStack> Stack_;
     TExceptionSafeContext MachineContext_;
 
     explicit TFiber(EExecutionStackKind stackKind = DefaultExecutionStackKind);
